+++ src/Aks/Aks/Commands/NewAzureRmAksNodePool.cs
@@ -131,6 +131,9 @@ namespace Microsoft.Azure.Commands.Aks
         [Parameter(Mandatory = false, HelpMessage = "The fully qualified resource ID of the Dedicated Host Group to provision virtual machines from, used only in creation scenario and not allowed to changed once set.")]
         public string HostGroupID { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "The ID of the subnet which pods will join when launched.")]
+        public string PodSubnetID { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -287,6 +290,9 @@ namespace Microsoft.Azure.Commands.Aks
             if (this.IsParameterBound(c => c.HostGroupID))
             {
                 agentPool.HostGroupID = HostGroupID;
+            }
+            if (this.IsParameterBound(c => c.PodSubnetID)) {
+                agentPool.PodSubnetID = PodSubnetID;
             }
 
             return agentPool;