+++ src/Aks/Aks/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added parameter `-NodePodSubnetID` for `New-AzAksCluster`, `-PodSubnetID` for `New-AzAksNodePool`
 * Fixed the issue that system variable `$IsWindows` is undefined in `Windows PowerShell`.
 * Decoupled AutoMapper dependency, replaced with AdapterHelper.